Move hashtable_mod_id32_dynamic implementation to it's own file
We will refactor that hashtable to use chunks of memory instead of one
contigious array. Extracting it to a file helps with code clarity
